requirements.txt
basicsr>=1.4.2
facexlib>=0.2.5
gfpgan>=1.3.5
numpy==1.23
opencv-python
pillow
flask
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
tqdm
	